Make the period for ktimer an unsigned int

+++ ktimer.h
@@ -5,14 +5,14 @@
 
 struct ktimer {
   int state;
-  long period_ms;
+  unsigned int period_ms;
   void (*callback)(void*);
   void *callback_arg;
   spinlock_t lock;
 };
 
 void ktimer_init(struct ktimer *t, void (*callback)(void*), void* arg);
-int ktimer_start(struct ktimer *t, int period_ms);
+int ktimer_start(struct ktimer *t, unsigned int period_ms);
 int ktimer_stop(struct ktimer *t);
 
 #endif // KTIMER_H
